+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 2.0
 
 * Removed methods deprecated in v1.9 (Issue #90, PR #92)
+* Fixed behavior of `-extractFilesTo:overwrite:error:`, so it shows the progress of each individual file as they extract (Issue #91, PR #94)
 
 ## 1.9